refactor(tests): extract matplotlib plotting to separate module
Move visualization utilities from test_demos.py to plot_utils.py:
- plot_paths(): Plot toolpaths with contours, holes, sharp corners
- save_plot(): Save figures to output directory
- plot_comparison(): Side-by-side comparison of original/resampled paths

This improves code organization and makes plotting utilities
reusable across different test modules.
